Moved Inbox Route to Auth.py
