+++ codebase/website/auth.py
@@ -3,11 +3,28 @@ from .models import User
 from werkzeug.security import generate_password_hash, check_password_hash
 from . import db   ##means from __init__.py import db
 from flask_login import login_user, login_required, logout_user, current_user
-
+from faker import Faker
 
 auth = Blueprint('auth', __name__)
 
+fake = Faker()
+def generate_emails(num_emails):
+    emails = []
+    for _ in range(num_emails):
+        email = {
+            'from': fake.email(),
+            'subject': fake.sentence(),
+            'body': fake.paragraph()
+        }
+        emails.append(email)
+    return emails
 
+@auth.route("/", methods=['GET', 'POST'])
+@login_required
+def inbox():
+    num_emails = 10
+    emails = generate_emails(num_emails)
+    return render_template("home.html", emails=emails, user=current_user)
 
 @auth.route('/login', methods=['GET', 'POST'])
 def login():
@@ -20,7 +37,7 @@ def login():
             if check_password_hash(user.password, password):
                 flash('Logged in successfully!', category='success')
                 login_user(user, remember=True)
-                return redirect(url_for('views.home'))
+                return redirect(url_for('auth.inbox'))
             else:
                 flash('Incorrect password, try again.', category='error')
         else:
